middleware/group-route/migration-issue

+++ database/migrations/2021_11_14_090650_article_tag.php
@@ -16,9 +16,9 @@ class ArticleTag extends Migration
         //
 
         Schema::create('article_tag', function (Blueprint $table) {
-        $table->integer('article_id')->unsigned()->nullable();
+        $table->unsignedBigInteger('article_id')->unsigned()->nullable();
         $table->foreign('article_id')->references('id')->on('articles')->onDelete('cascade');
-        $table->integer('tag_id')->unsigned()->nullable();
+        $table->unsignedBigInteger('tag_id')->unsigned()->nullable();
         $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
         });
     }
